+++ src/types/zodAffect.ts
@@ -7,6 +7,7 @@ import {
 
 import { zodOsimDateTime, ImpactEnumWithBlank, ZodFlawClassification, ZodAlertSchema } from './zodShared';
 import { omit, pick } from 'ramda';
+import type { ValueOf } from '@/utils/typeHelpers';
 
 
 const AffectednessEnumWithBlank = { 'Empty': '', ...AffectednessEnum } as const;
@@ -16,11 +17,22 @@ export const affectImpacts = { ...omit([''], ImpactEnumWithBlank), Empty: '' } a
 export const affectResolutions = ResolutionEnumWithBlank;
 export const affectAffectedness = AffectednessEnumWithBlank;
 
-export const AffectednessResolutionPairs: Record<AffectednessEnum, Partial<typeof ResolutionEnumWithBlank>> = {
-  AFFECTED: pick(['Delegated', 'Wontfix', 'Ooss'], affectResolutions),
-  NEW: pick(['Empty', 'Wontfix', 'Ooss'], affectResolutions),
-  NOTAFFECTED: pick(['Empty'], affectResolutions),
-};
+type PossibleResolutions = Record<AffectednessEnum, Partial<typeof ResolutionEnumWithBlank>>;
+
+export function possibleAffectResolutions(impact?: ValueOf<typeof affectImpacts> | null): PossibleResolutions {
+  const pickResolution = (options: (keyof typeof ResolutionEnumWithBlank)[], resolutions: typeof affectResolutions) => {
+    const possibleResolutions = options.filter(option => (
+      impact === affectImpacts.Low && option === 'Defer' || option !== 'Defer'
+    ));
+    return pick(possibleResolutions, resolutions);
+  };
+
+  return {
+    AFFECTED: pickResolution(['Delegated', 'Defer', 'Ooss'], affectResolutions),
+    NEW: pickResolution(['Empty', 'Defer', 'Ooss'], affectResolutions),
+    NOTAFFECTED: pickResolution(['Empty'], affectResolutions),
+  };
+}
 
 
 export type ErratumSchemaType = typeof ErratumSchema;